Git ReviewBoard Integration
Summary: In this commit I have added automatic support for code review. As soon
as you commit something, it is submitted to our ReviewBoard server.

This is how it is currently set-up:

`commit-msg` is just a link a to `ReviewBoard.py`.
`ReviewBoard.py` takes one parameter: a file that holds commit message. It then
modifies the file acordingly and submits the contents to the server.
`ReviewBoardInvoker.py` and `CommitMessage.py` just hold classes that are used
in `ReviewBoard.py`.

Tags: git, git-hooks

Reviewers: agrbin, ivankovic

Groups:

Blamed Revisions:

Tickets:

Skip Reviewboard: yes

ReviewBoardID: 21