Haski - tool for code review & much more
summary: This is a new tool for code review and much more. The code is awful,
but it gets shit done. I am in no way happy with it.

This is the suggested workflow:

1. work on your commit
2. commit it localy
3. `./haski lint` -> lint the code
4. if errors, warnings: go to step 1
5. `./haski review` -> sends the code to reviewboard
6. if bugs, comments: go to step 1.
7. push the code globally

Haski currently supports the following:
- 'lint': lints the code
- 'review': sends the code to reviewboard
- 'license': prints the license from COPYING in the root dir of the project

Known bugs
If you have something staged but not yet commited, please note that the script
will add it the commit. Will look into fixing it.

Other (minor) changes
- moved scripts from bin folder to scripts/bin folder
- changed the commit.template
- removed git hooks

reviewers: agrbin, prasko, ivankovic

reviewboard id: 38