[Haski] Refactoring + attaching lint results
summary: I finally came around to refactoring the linting code. Now you
can easily add new linters. Besides that, linting is now performed on
the files that are in the repository in the specified commit and not on
the ones that you are currently working on.

To add a new linter, you now have to write a class that extends
baselinter and implements two functions: _filter_files() and lint().
The former takes a list of file names and returns the files that the
class can lint, and the latter invokes the linting command. All you
have to do then is to add the linter's class name to the LINTERS list
in commit.py.

Another addition is an option to skip verifying the commit message
when performing the lint task.

reviewers: agrbin, ivankovic, prasko

tickets: 2, 3

reviewboard id: 49