[Haski] Bugfixes
summary: ReviewBoard Invoker had a bug - it would try to print a
variable that doesn't exist in case of post-review failure. That, of
course causes a the Invoker itself to crash. Hopefully, this should fix
it. I will have to re-write the invoker some day before it becomes too
ugly.

reviewers: agrbin, ivankovic, prasko

reviewboard id: 52